back/requirements.txt
Flask
Flask-SQLAlchemy
python-decouple
psycopg2-binary